requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
MarkupSafe==0.23
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.10.0
six==1.10.0
Werkzeug==0.11.10
